Private and public cards

+++ app/api/cards.py
@@ -19,43 +19,63 @@ cardsRouter = APIRouter(prefix="/cards")
 
 
 @cardsRouter.get("/getAllCards/", response_model=List[CardSchema])
-def get_all_cards(current_user: Annotated[UserSchema, Depends(get_current_active_user)]):
+def get_all_cards_by_user(author_id, current_user: Annotated[UserSchema, Depends(get_current_active_user)]):
+    """ Возвращает все свои/чужие карточки """
     repository_cards = RepositoryCards(session=config_session_maker())
+    if author_id == current_user.id:
+        cards = repository_cards.get_all_cards_by_user(author_id=author_id)
+    else:
+        cards = repository_cards.get_all_public_cards(author_id=author_id)
 
-    return repository_cards.get_all_cards(user_id=current_user.id)
+    return cards
 
 
 @cardsRouter.post("/addCard/", response_model=CardSchema)
 def add_card(card: CardSchema, current_user: Annotated[UserSchema, Depends(get_current_active_user)]):
-
+    """Добавляет карточку"""
     repository_cards = RepositoryCards(session=config_session_maker())
     if repository_cards.get_card_by_id(card.id):
         raise HTTPException(status_code=403, detail="Запись с таким id уже существует")
+    if card.user_fk != current_user.id:
+        raise HTTPException(status_code=403, detail="Нельзя добавлять карточки чужим пользователям")
     repository_cards.add_card(card)
     return card
 
 
 @cardsRouter.put("/change_card")
-def change_card(card: CardSchema):
+def change_card(card: CardSchema, current_user: Annotated[UserSchema, Depends(get_current_active_user)]):
+    """ Изменяет карточку """
+    if card.user_fk != current_user.id:
+        raise HTTPException(status_code=403, detail="Пользователь не может менять чужие карточки")
     repository_card = RepositoryCards(session=config_session_maker())
     repository_card.change_card(card)
     return card
 
 @cardsRouter.post("/delete_card")
-def delete_card(card_id):
+def delete_card(card_id, current_user: Annotated[UserSchema, Depends(get_current_active_user)]):
+    """ Удаляет карточку """
     repository_card = RepositoryCards(session=config_session_maker())
+    card = repository_card.get_card_by_id(card_id=card_id)
+    # with open("log.txt", "w", encoding="utf-8") as file:
+    #     file.write(f"{card.user_fk}, {current_user.id}, \n")
+    if card.user_fk != current_user.id:
+        raise HTTPException(status_code=403, detail="Пользователь не может удалять чужие карточки")
     repository_card.delete_card(card_id)
     return f"{card_id=} карточка удалена"
 
 @cardsRouter.post("/solve_card")
 def solve_card(solve_card: SolveCardSchema,
                current_user: Annotated[UserSchema, Depends(get_current_active_user)]):
+    """ Добавляет решение карточки """
     repository_solve_card = RepositorySolveCards(session=config_session_maker())
     repository_solve_card.add_solve_card(solve_card)
     return solve_card
 
 @cardsRouter.post("/get_solves_by_user")
-def solve_card(user_id: int,
-               current_user: Annotated[UserSchema, Depends(get_current_active_user)]):
+def get_solves_by_user(current_user: Annotated[UserSchema, Depends(get_current_active_user)]):
+    """ возвращает все решения авторизованного пользователя """
+
     repository_solve_card = RepositorySolveCards(session=config_session_maker())
-    return repository_solve_card.get_solves_by_user(user_id)
+    return repository_solve_card.get_solves_by_user(current_user.id)
+
+
